+++ packages/lit-virtualizer/CHANGELOG.md
@@ -3,5 +3,8 @@
 - `scrollToIndex` method on `<lit-virtualizer>`.
 - `scrollToIndex` configuration option on `scroll` directive.
 
+### Changed
+- Renamed API option `template` to `renderItem`.
+
 ## [0.1.0] - 2019-06-05
 - Initial prerelease.